Relayminer tests are passing again
